% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ado.ano.plot.R
\name{ado.ano.plot}
\alias{ado.ano.plot}
\title{plot the results of pairwise.adonis and pairwise.anosim.}
\usage{
ado.ano.plot(data, limits, midpoint, show = 1)
}
\arguments{
\item{data}{a data.frame from pairwise.adonis or pairwise.anosim}

\item{limits}{range of heatmap}

\item{midpoint}{midpoint of heatmap}

\item{show}{return the plot of heatmap or the p.value pointmap. 1 for heatmap
2 for p.value pointmap, the default value is 1.}
}
\value{
as describe in show
}
\description{
plot the results of pairwise.adonis and pairwise.anosim.
}
\examples{
ado.ano.plot(data = adonis.pairwise(iris[,1:4],iris$Species),
limit = c(0, 1), midpoint = 0.5, show = 1)
}
\author{
ZH Gai
}
