% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis.pairwise.R
\name{adonis.pairwise}
\alias{adonis.pairwise}
\title{Pairwise multilevel comparison using adonis}
\usage{
adonis.pairwise(
  x,
  factors,
  sim.function = "vegdist",
  sim.method = "bray",
  p.adjust.m = "bonferroni",
  perm = 999
)
}
\arguments{
\item{x}{Data frame (the community table), or "dist" object (user-supplied distance matrix).}

\item{factors}{Vector (a column or vector with the levels to be compared pairwise).}

\item{sim.function}{Function used to calculate the similarity matrix,
one of 'daisy' or 'vegdist' default is 'vegdist'. Ignored if x is a distance matrix.}

\item{sim.method}{Similarity method from daisy or vegdist, default is 'bray'. Ignored if x is a distance matrix.}

\item{p.adjust.m}{The p.value correction method, one of the methods supported by p.adjust(),
default is 'bonferroni'.}

\item{perm}{perm}
}
\value{
Table with the pairwise factors, R^2, p.value and adjusted p.value.
}
\description{
Pairwise multilevel comparison using adonis
}
\examples{
adonis.pairwise(iris[,1:4],iris$Species)
}
\author{
ZH Gai
}
