% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Calculate the alpha diversity of OUT table or ASVs table using estimateR and diversity function of vegan package.}
\usage{
alpha(data)
}
\arguments{
\item{data}{a data frame containing the grouping information}
}
\value{
a data frame containg the alpha index
}
\description{
Calculate the alpha diversity of OUT table or ASVs table using estimateR and diversity function of vegan package.
}
\examples{
library(vegan)
data(BCI)
BCI <- data.frame(group = c(rep("ss", 25), rep("dd", 25)), BCI)
BCI$group <- as.factor(BCI$group)
alpha(BCI)  |> ggroup::ggboxplot.LetSig()
}
\author{
ZHonghui Gai
}
