% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anosim.pairwise.R
\name{anosim.pairwise}
\alias{anosim.pairwise}
\title{Pairwise multilevel comparison using anosim}
\usage{
anosim.pairwise(
  x,
  grouping,
  sim.function = "vegdist",
  sim.method = "bray",
  p.adjust.m = "bonferroni",
  perm = 999
)
}
\arguments{
\item{x}{Data matrix or data frame in which rows are samples and columns are response variable(s),
or a dissimilarity object or a symmetric square matrix of dissimilarities}

\item{grouping}{Factor for grouping observations.}

\item{sim.function}{Function used to calculate the similarity matrix,
one of 'daisy' or 'vegdist' default is 'vegdist'. Ignored if x is a distance matrix.}

\item{sim.method}{Similarity method from daisy or vegdist, default is 'bray'. Ignored if x is a distance matrix.}

\item{p.adjust.m}{The p.value correction method, one of the methods supported by p.adjust(),
default is 'bonferroni'.}

\item{perm}{permutations}
}
\value{
Table with the pairwise factors, anosim R, p.value and adjusted p.value.
}
\description{
Pairwise multilevel comparison using anosim
}
\examples{
anosim.pairwise(iris[, 1:4], iris$Species)
}
\author{
ZH Gai
}
