% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.size.adonis.R
\name{effect.size.adonis}
\alias{effect.size.adonis}
\title{The adonis function in vegan package is used to analyze the effect of each variable of one data frame on another data matrix.
It can be metadata such as age and gender, the effect on the gut microbiota.
It can also be each OTU in the gut microbiota, the effect on the distance matrix of inflammatory factors.}
\usage{
effect.size.adonis(data1, data2, signif = TRUE, barplot = TRUE)
}
\arguments{
\item{data1}{a matrix, dependent variable, equivalent to y of y ~x}

\item{data2}{a data frame, independent variable, equivalent to x of y ~x}

\item{signif}{select the significant variables}

\item{barplot}{show the barplot}
}
\value{
a ggplot2 function
}
\description{
The adonis function in vegan package is used to analyze the effect of each variable of one data frame on another data matrix.
It can be metadata such as age and gender, the effect on the gut microbiota.
It can also be each OTU in the gut microbiota, the effect on the distance matrix of inflammatory factors.
}
\examples{
effect.size.adonis(iris[, 1:2], iris[, 3:5, drop = F])
}
\author{
Zhonghui Gai
}
