% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggadonis.table.R
\name{ggadonis.table}
\alias{ggadonis.table}
\title{Transform data frame to a ggplot object table, which can be inserted to a ggplot graph}
\usage{
ggadonis.table(table)
}
\arguments{
\item{table}{a data frame table}
}
\value{
a ggplot2 object
}
\description{
Transform data frame to a ggplot object table, which can be inserted to a ggplot graph
}
\examples{
g <- ggadonis.table(table = adonis.pairwise(iris[,1:4],iris$Species))
ggplot(data = data.frame(x = 1:10, y = 1:10), aes(x,y)) + geom_point() +
theme_bw() +
  annotation_custom(grob = g,
                    xmin = 3,xmax = 8,
                    ymin = 8,ymax = 8.5)
}
\author{
Zhonghui Gai
}
