% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggadonis.table.R
\name{ggadonis.table_plot}
\alias{ggadonis.table_plot}
\title{Place the table in a separate figure.}
\usage{
ggadonis.table_plot(gtable)
}
\arguments{
\item{gtable}{the result of ggadonis.table function}
}
\value{
a ggplot2 object
}
\description{
Place the table in a separate figure.
}
\examples{
adonis.pairwise(iris[,1:4],iris$Species) |>
ggadonis.table() |>
  ggadonis.table_plot()
}
\author{
Zhonghui Gai
}
