% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbeta.disp.R
\name{ggbeta.disp}
\alias{ggbeta.disp}
\title{Multivariate dispersion analysis based on ggplot2 and betadisper function of vagan package}
\usage{
ggbeta.disp(data, type = "ordiplot", pc = 12, show.pvalue = TRUE)
}
\arguments{
\item{data}{a data frame with the first collounm as the grouping information}

\item{type}{the type of result to present, one of "boxplot", "ordiplot" or "mod", default value is ordiplot.}

\item{pc}{which axis to present, one of 12, 13, or 23, the default value is 12.}

\item{show.pvalue}{a boolean value, the default value is TRUE.}
}
\value{

}
\description{
Multivariate dispersion analysis based on ggplot2 and betadisper function of vagan package
}
\examples{
genus.Syn <- read.csv(file = "genus.Syn.csv", row.names = 1)
ggbeta.disp(genus.Syn, pc = 12)
library(vegan)
plot(mod, ellipse = T, hull = F, conf = 0.95, axes = c(1,2))
vegan:::plot.betadisper(mod)
vegan:::boxplot.betadisper(mod)
ggbeta.disp(genus.Syn, pc = 12, type = "boxplot")
ggbeta.disp(genus.Syn, pc = 12, type = "mod")
}
\author{
Zhonghui Gai
}
